requirements.txt
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.11.0
gunicorn==20.1.0
Pillow==8.3.0
PyJWT==2.3.0
pylint==2.6.0
sqlparse==0.4.1
whitenoise==6.0.0
psycopg2-binary==2.9.3
pytz==2021.1